+++ Changelog.md
@@ -13,6 +13,8 @@
 - test: add tests for malformed inputs for serde deserializer
 - fix: allow to deserialize `unit`s from any data in attribute values and text nodes
 - refactor: unify errors when EOF encountered during serde deserialization
+- test: ensure that after deserializing all XML was consumed
+- feat: add `Deserializer::from_str` and `Deserializer::from_bytes`
 
 ## 0.23.0-alpha3